Extract task signing helper functions
