+++ helpers/test-helper.js
@@ -2,10 +2,8 @@
 /* eslint-disable no-console */
 import shortid from "shortid";
 import chai from "chai";
-import { asciiToHex, sha3, soliditySha3, padLeft } from "web3-utils";
-import { hashPersonalMessage, ecsign } from "ethereumjs-util";
+import { asciiToHex } from "web3-utils";
 import BN from "bn.js";
-import fs from "fs";
 
 import { UINT256_MAX, MIN_STAKE, MINING_CYCLE_DURATION, DEFAULT_STAKE } from "./constants";
 
@@ -302,70 +300,6 @@ export async function forwardTime(seconds, test) {
     }
   });
   return p;
-}
-
-export function getFunctionSignature(sig) {
-  return sha3(sig).slice(0, 10);
-}
-
-export async function createSignatures(colony, taskId, signers, value, data) {
-  const sourceAddress = colony.address;
-  const destinationAddress = colony.address;
-  const nonce = await colony.getTaskChangeNonce(taskId);
-  const accountsJson = JSON.parse(fs.readFileSync("./ganache-accounts.json", "utf8"));
-
-  const input = `0x${sourceAddress.slice(2)}${destinationAddress.slice(2)}${padLeft(value.toString(16), "64", "0")}${data.slice(2)}${padLeft(
-    nonce.toString(16),
-    "64",
-    "0"
-  )}`; // eslint-disable-line max-len
-  const sigV = [];
-  const sigR = [];
-  const sigS = [];
-  const msgHash = soliditySha3(input);
-
-  for (let i = 0; i < signers.length; i += 1) {
-    let user = signers[i].toString();
-    user = user.toLowerCase();
-    const privKey = accountsJson.private_keys[user];
-    const prefixedMessageHash = await hashPersonalMessage(Buffer.from(msgHash.slice(2), "hex"));
-    const sig = await ecsign(prefixedMessageHash, Buffer.from(privKey, "hex"));
-
-    sigV.push(sig.v);
-    sigR.push(`0x${sig.r.toString("hex")}`);
-    sigS.push(`0x${sig.s.toString("hex")}`);
-  }
-
-  return { sigV, sigR, sigS };
-}
-
-export async function createSignaturesTrezor(colony, taskId, signers, value, data) {
-  const sourceAddress = colony.address;
-  const destinationAddress = colony.address;
-  const nonce = await colony.getTaskChangeNonce(taskId);
-  const accountsJson = JSON.parse(fs.readFileSync("./ganache-accounts.json", "utf8"));
-  const input = `0x${sourceAddress.slice(2)}${destinationAddress.slice(2)}${padLeft(value.toString(16), "64", "0")}${data.slice(2)}${padLeft(
-    nonce.toString(16),
-    "64",
-    "0"
-  )}`; // eslint-disable-line max-len
-  const sigV = [];
-  const sigR = [];
-  const sigS = [];
-  const msgHash = soliditySha3(input);
-
-  for (let i = 0; i < signers.length; i += 1) {
-    let user = signers[i].toString();
-    user = user.toLowerCase();
-    const privKey = accountsJson.private_keys[user];
-    const prefixedMessageHash = soliditySha3("\x19Ethereum Signed Message:\n\x20", msgHash);
-    const sig = ecsign(Buffer.from(prefixedMessageHash.slice(2), "hex"), Buffer.from(privKey, "hex"));
-    sigV.push(sig.v);
-    sigR.push(`0x${sig.r.toString("hex")}`);
-    sigS.push(`0x${sig.s.toString("hex")}`);
-  }
-
-  return { sigV, sigR, sigS };
 }
 
 export function bnSqrt(bn, isGreater) {
